requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.1
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
aiohappyeyeballs=2.6.1=pyhd8ed1ab_0
aiohttp=3.11.18=py312h178313f_0
aiosignal=1.3.2=pyhd8ed1ab_0
attrs=25.3.0=pyh71513ae_0
brotli-python=1.1.0=py312h2ec8cdc_2
bzip2=1.0.8=h4bc722e_7
ca-certificates=2025.4.26=hbd8a1cb_0
cachetools=5.5.2=pyhd8ed1ab_0
certifi=2025.4.26=pyhd8ed1ab_0
cffi=1.17.1=py312h06ac9bb_0
charset-normalizer=3.4.2=pyhd8ed1ab_0
click=8.2.0=pyh707e725_0
colorama=0.4.6=pyhd8ed1ab_1
cryptography=44.0.3=py312hda17c39_0
cyrus-sasl=2.1.27=h54b06d7_7
docker-compose=2.36.0=ha770c72_0
frozenlist=1.6.0=py312hb9e946c_0
google-api-core=2.24.2=pyhd8ed1ab_0
google-api-python-client=2.169.0=pyhff2d567_0
google-auth=2.40.1=pyhd8ed1ab_0
google-auth-httplib2=0.2.0=pyhd8ed1ab_1
googleapis-common-protos=1.70.0=pyhd8ed1ab_0
greenlet=3.2.2=py312h2ec8cdc_0
h2=4.2.0=pyhd8ed1ab_0
hpack=4.1.0=pyhd8ed1ab_0
httplib2=0.22.0=pyhd8ed1ab_1
hyperframe=6.1.0=pyhd8ed1ab_0
icu=75.1=he02047a_0
idna=3.10=pyhd8ed1ab_1
isodate=0.7.2=pyhd8ed1ab_1
joblib=1.5.0=pyhd8ed1ab_0
keyutils=1.6.1=h166bdaf_0
krb5=1.21.3=h659f571_0
ld_impl_linux-64=2.43=h712a8e2_4
libabseil=20250127.1=cxx17_hbbce691_0
libblas=3.9.0=31_h59b9bed_openblas
libcblas=3.9.0=31_he106b2a_openblas
libedit=3.1.20250104=pl5321h7949ede_0
libexpat=2.7.0=h5888daf_0
libffi=3.4.6=h2dba641_1
libgcc=15.1.0=h767d61c_2
libgcc-ng=15.1.0=h69a702a_2
libgfortran=15.1.0=h69a702a_2
libgfortran5=15.1.0=hcea5267_2
libgomp=15.1.0=h767d61c_2
liblapack=3.9.0=31_h7ac8fdf_openblas
liblzma=5.8.1=hb9d3cd8_1
libnsl=2.0.1=hd590300_0
libntlm=1.8=hb9d3cd8_0
libopenblas=0.3.29=pthreads_h94d23a6_0
libpq=17.5=h27ae623_0
libsqlite=3.49.2=hee588c1_0
libstdcxx=15.1.0=h8f9b012_2
libstdcxx-ng=15.1.0=h4852527_2
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=hb9d3cd8_2
multidict=6.4.3=py312h178313f_0
ncurses=6.5=h2d0b736_3
nltk=3.9.1=pyhd8ed1ab_1
numpy=2.2.5=py312h72c5963_0
openldap=2.6.9=he970967_0
openssl=3.5.0=h7b32b05_1
pandas=2.2.3=py312hf9745cd_3
pip=25.1.1=pyh8b19718_0
propcache=0.3.1=py312h178313f_0
proto-plus=1.26.1=pyhd8ed1ab_0
protobuf=5.29.3=py312h0f4f066_0
psycopg2=2.9.9=py312hfaedaf9_2
pyasn1=0.6.1=pyhd8ed1ab_2
pyasn1-modules=0.4.2=pyhd8ed1ab_0
pycparser=2.22=pyh29332c3_1
pyopenssl=25.0.0=pyhd8ed1ab_0
pyparsing=3.2.3=pyhd8ed1ab_1
pysocks=1.7.1=pyha55dd90_7
python=3.12.10=h9e4cc4f_0_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python-dotenv=1.1.0=pyh29332c3_1
python-tzdata=2025.2=pyhd8ed1ab_0
python_abi=3.12=7_cp312
pytz=2025.2=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_1
readline=8.2=h8c095d6_2
regex=2024.11.6=py312h66e93f0_0
requests=2.32.3=pyhd8ed1ab_1
rsa=4.9.1=pyhd8ed1ab_0
setuptools=80.1.0=pyhff2d567_0
six=1.17.0=pyhd8ed1ab_0
sqlalchemy=2.0.40=py312h66e93f0_0
tk=8.6.13=noxft_h4845f30_101
tqdm=4.67.1=pyhd8ed1ab_1
typing-extensions=4.13.2=h0e9735f_0
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025b=h78e105d_0
uritemplate=4.1.1=pyhd8ed1ab_1
urllib3=2.4.0=pyhd8ed1ab_0
wheel=0.45.1=pyhd8ed1ab_1
yarl=1.20.0=py312h178313f_0
zstandard=0.23.0=py312h66e93f0_2
